use default font size on Android
This solves the root problem of most of the dpi-depdent sizing in the
mobile Android version.

When setting a custom font size, we circumvent Android's defaults, which
means that we end up setting the font size further down the road to get
readable fonts.

I suppose this was set in order to make the QWidget-based user interface
"work" on Android. Hard-coding a font size at this central point brings
more headache than needed, one could instead consider adding an #ifdef
Q_OS_ANDROID in main window.

Signed-off-by: Sebastian Kügler <sebas@kde.org>