Recognize Android as CMAKE_SYSTEM_NAME
This fixes the build for me against Qt 5.5. It seems Android is not
Linux anymore, but uses its own CMAKE_SYSTEM_NAME now, so in this case
we also want to compile android.cpp.

This change should be rather safe, since it catches a condition that was
not previously handled.

Signed-off-by: Sebastian Kügler <sebas@kde.org>